fix: Move functions
