+++ main/parseCommand.go
@@ -3,34 +3,12 @@ package main
 import (
 	"errors"
 	"fmt"
-	"log"
 	"regexp"
 	"strings"
-
-	"github.com/bwmarrin/discordgo"
 )
 
-// OnMessageCreate is called when there is a new message in a guild this bot is belogns to.
-// If this bot is mentioned, parse command and do corresponding actions.
-func OnMessageCreate(session *discordgo.Session, message *discordgo.MessageCreate) {
-	if message.Author.ID == session.State.User.ID || len(message.Mentions) == 0 || message.Mentions[0].Username != session.State.User.Username {
-		return
-	}
-
-	command, err := ParseCommand(message.Content)
-	if err != nil {
-		session.ChannelMessageSend(message.ChannelID, err.Error())
-		log.Println(err)
-		return
-	}
-
-	if err = command.handle(session, message.Message); err != nil {
-		session.ChannelMessageSend(message.ChannelID, err.Error())
-		log.Println(err)
-	}
-}
-
 // ParseCommand parses messages from Discord and returns results as sturct.
+// Assumes first word of the input as a mention to this bot.
 func ParseCommand(input string) (Command, error) {
 	arguments := strings.Fields(input)
 	if len(arguments) < 2 {
